Check for null during node process disposal. Fixes #1061

+++ src/Microsoft.AspNetCore.NodeServices/HostingModels/OutOfProcessNodeInstance.cs
@@ -297,7 +297,7 @@ namespace Microsoft.AspNetCore.NodeServices.HostingModels
 
                 // Make sure the Node process is finished
                 // TODO: Is there a more graceful way to end it? Or does this still let it perform any cleanup?
-                if (!_nodeProcess.HasExited)
+                if (_nodeProcess != null && !_nodeProcess.HasExited)
                 {
                     _nodeProcess.Kill();
                 }
